(ch1/1-2-4): cleaned up widget based code
