+++ Chap1-Events/1-2EventsDemo/main.py
@@ -1,13 +1,36 @@
 import sys
-from PySide6.QtWidgets import QApplication
-from widget import Widget
+from pathlib import Path
 
-def main():
-    app = QApplication(sys.argv)
-    window = Widget()
-    window.show()
+from PySide6.QtGui import QGuiApplication
+from PySide6.QtQml import QQmlApplicationEngine
+from PySide6.QtCore import QObject, Slot
+
+
+class EventLogger(QObject):
+    """Python backend class to handle logging events from QML"""
     
-    return app.exec()
+    @Slot(str)
+    def log(self, message):
+        """Log messages from QML to console"""
+        print(message)
+
 
 if __name__ == "__main__":
-    sys.exit(main())
+    app = QGuiApplication(sys.argv)
+    
+    # Create QML engine
+    engine = QQmlApplicationEngine()
+    
+    # Create and expose the EventLogger to QML
+    logger = EventLogger()
+    engine.rootContext().setContextProperty("eventLogger", logger)
+    
+    # Load QML file
+    qml_file = Path(__file__).resolve().parent / "main.qml"
+    engine.load(qml_file)
+    
+    # Check if loading was successful
+    if not engine.rootObjects():
+        sys.exit(-1)
+        
+    sys.exit(app.exec())
