--- Chap1-Events/1-2EventsDemo/widget.ui
@@ -1,20 +1,0 @@
-<ui version="4.0">
- <class>Widget</class>
- <widget class="QWidget" name="Widget" >
-  <property name="geometry" >
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>400</width>
-    <height>300</height>
-   </rect>
-  </property>
-  <property name="windowTitle" >
-   <string>Widget</string>
-  </property>
- </widget>
- <layoutDefault spacing="6" margin="11" />
- <pixmapfunction></pixmapfunction>
- <resources/>
- <connections/>
-</ui>
